requirements.txt
aiogram==2.23.1
openai==0.28
python-dotenv






